create applytransaction.go
